"Display alert for location, get weather forecast using city name from user's current location"
This commit contains a modification to the existing code that displays an alert prompting the user to turn on their location. Additionally, it fetches the user's current location's latitude and longitude and uses them to retrieve the city name. The city name is then used to get the weather forecast from the OpenWeatherMap API.